Update PipeNet (primarily resource processor)
Non-resource processor changes:

- `CompRefillWithPipes` and `PipeNet:RefillableAmount` now consider the refueling cost from difficulty
- Several `PipeNet` methods that distribute resources across pipe net now accept an (optional) argument if they should also draw resource from storages
- `Process` when distributing to a non-storage source won't draw from the storage tanks
- `CompResource` will only spawn and maintain the sustainer while spawned (better minified handling)

`CompResourceProcessor` changes:

- When outputting resources, it'll distribute as much as possible (and keep excess to distribute later) rather than doing nothing
- Resource processor now supports fractional input and output of resources
  - In case of things, the value will be rounded down and added to the next output (added to excess output)
- When distributing, other resource processors may be filled
- Improved handling when minified
- Fixed multiple enumeration issue from `CompGetGizmosExtra` method
- Fixed resources being wasted trying to refill (almost) full refuelables
- Fixed issue where resource processor wouldn't be treated as not working if it's not connected to a pipe network and can't distribute as items
- Fixed an edge case scenario of distributing to refuelables/converters potentially using resources from storages
- Fixed a potential edge case scenario where resource processors (requiring a very specific resource count) wouldn't work due to floating point imprecision

+++ Source/PipeSystem/PipeSystem/Comps/CompProperties/CompProperties_ResourceProcessor.cs
@@ -38,13 +38,14 @@ namespace PipeSystem
             // Produce each ticks: Default to 600 ticks (10 sec)
             public int eachTicks = 600;
             // Amount needed to produce result
-            public int countNeeded;
+            public float countNeeded;
             // Result as a thing
             public ThingDef thing;
-            public int thingCount;
+            // Count will be rounded down to a while number, and any fractional reminders kept for further processes
+            public float thingCount;
             // Result as a piped resource
             public PipeNetDef net;
-            public int netCount;
+            public float netCount;
         }
     }
 }
